refactor: decompose binGenerator.ts into focused modules
Split the 1826-line monolith into single-responsibility modules while
preserving all existing behavior and public API:

- generatorTypes.ts: shared types, constants, utility functions
- shapeCache.ts: single-entry shape caches with getter/setter API
- socketBuilder.ts: socket cell generation + magnet/screw holes
- boxBuilder.ts: box body extrusion + stacking lip sweep
- featureBuilder.ts: compartments, inserts, cutouts, labels, scoops
- index.ts: barrel re-exporting the public API

binGenerator.ts is now a 560-line orchestrator composing the pipeline.
Each new module has colocated tests (55 new tests). All 148 existing
tests pass including 71 geometry snapshot tests confirming identical output.
